requirements.txt
# GUI framework
PyQt5>=5.15.0
PyTaskbar>=0.1.0

# Image processing
Pillow>=9.0.0
imagehash>=4.3.0

# Archive processing
rarfile>=4.0

# Data processing
numpy>=1.21.0

# Config
PyYAML>=6.0.0

# System operations
natsort>=8.3.0
psutil>=5.9.0
send2trash>=1.8.0
pywin32>=303

# Logging
loguru>=0.7.0
